+++ src/test/java/Ders1/HomeWork.java
@@ -1,3 +1,5 @@
+package Ders1;
+
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
