+++ src/test/java/Ders1/driverTanimlama.java
@@ -1,3 +1,5 @@
+package Ders1;
+
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -5,29 +7,35 @@ import org.testng.annotations.Test;
 
 import java.util.concurrent.TimeUnit;
 
-public class ilkKomutlar {
+
+public class driverTanimlama {
 
     static WebDriverManager driver;
 
-
     @Test
-    public void driver_Tanimlama() throws InterruptedException {
 
+    public void driver_Tanimlama(){
+// Driverimizi kullanilabilir hale getirdik
         WebDriver driver;
-
+// Driverimizi chromedrivera tanitiyoruz
         WebDriverManager.chromedriver().setup();
+//driver nesnesi olusturuyoruz
         driver=new ChromeDriver();
+//sayfamizi tam ekran yapiyoruz
         driver.manage().window().maximize();
-        driver.manage().timeouts().implicitlyWait(10 , TimeUnit.SECONDS);
+//sayfamizin yuklenmesi icin driverimizi bekletiyoruz
+        driver.manage().timeouts().implicitlyWait(10,TimeUnit.SECONDS);
+
+        //ilk deneme
         driver.get("https://google.com");
 
-        Thread.sleep(3000);
-
-        //calistigimiz pencereyi kapatir
-        driver.close();
-
-        //calistigimiz tum pencereleri kapatir.
-          // driver.quit();
-
     }
+
+
+
+
+
+
+
+
 }
