+++ src/test/java/Ders2/Id_name.java
@@ -1,0 +1,23 @@
+package Ders2;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.Test;
+
+import java.util.concurrent.TimeUnit;
+
+public class Id_name {
+
+    static WebDriver driver;
+
+    @Test
+    public void driverTanimla() {
+        WebDriver driver;
+        WebDriverManager.chromedriver().setup();
+        driver = new ChromeDriver();
+        driver.manage().window().maximize();
+        driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+
+        driver.get("http://gcreddy.com/project/admin/login.php");
+    }
+}
